move euth/IdeaColl/Workbench Module to euth/IdeaColl
